+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Unreleased
 
 - Added pager (if the text to display is long, you can scroll).
+- Added `configure clear` command.
 
 ## 0.7.1